fix borders
